Histogram and Summary expire individual entries
Histogram and Summary types now track the AddTime at the individual
entry level (Sum, Count, Quantile, and Bucket) and then expire those
entires based off that time.

Histogram and Summary still track the latest AddTime at the root of
the struct and use that time to expire the entire metric.